% Auto-generated: do not edit by hand
\name{dashCollapsibleTree}

\alias{dashCollapsibleTree}

\title{DashCollapsibleTree component}

\description{

}

\usage{
dashCollapsibleTree(id=NULL, items=NULL, topIds=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks}

\item{items}{Named list. }

\item{topIds}{Unnamed list. }
}
