% Auto-generated: do not edit by hand
\name{dashCollapsibleTree}

\alias{dashCollapsibleTree}

\title{DashCollapsibleTree component}

\description{

}

\usage{
dashCollapsibleTree(id=NULL, data=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks}

\item{data}{Unnamed list. }
}
